@@ -2,10 +2,10 @@ class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
-  before_filter :mobile_filter_header
+  before_action :mobile_filter_header
 
   def mobile_filter_header
-    @mobile = true 
+    @mobile = true
   end
 
 end
