@@ -1,5 +1,5 @@
 ActiveAdmin.register Restroom do
 
-permit_params :name, :street, :city, :state, :accessible, :unisex, :directions,
+permit_params :name, :street, :city, :state, :accessible, :changing_table, :unisex, :directions,
               :comment, :latitude, :longitude, :country
 end
